[issue-221] Moved files
