+++ cocoon/frontend/src/common/csrftoken.js
