+++ cocoon/frontend/src/common/homeTile/homeTile.js
